AVRO-4139 [java] equality check for maps in arrays (#3378)
With the introduction of the `AbstractArray` in avro 1.12, it was not
possible anymore to compare arrays which contained objects which
contained a Map, as the fact that we compared for equality was not
propagated anymore.

This commit re-introduces this propagation.

(cherry picked from commit a651bb2f5d8ad87e17fc66344a617d822b3611f3)